refactor: Extract code that creates window.Vaadin.views json
