+++ packages/java/endpoint/src/main/java/com/vaadin/hilla/route/RouteUnifyingIndexHtmlRequestListener.java
@@ -35,6 +35,8 @@ import com.vaadin.flow.server.menu.AvailableViewInfo;
 import com.vaadin.flow.server.menu.MenuRegistry;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.core.JsonProcessingException;
+
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -63,55 +65,20 @@ public class RouteUnifyingIndexHtmlRequestListener
 
     private static final Logger LOGGER = LoggerFactory
             .getLogger(RouteUnifyingIndexHtmlRequestListener.class);
-    private final NavigationAccessControl accessControl;
-    private final DeploymentConfiguration deploymentConfiguration;
-    private final boolean exposeServerRoutesToClient;
-    private final ObjectMapper mapper = new ObjectMapper();
-    private final ViewAccessChecker viewAccessChecker;
 
-    /**
-     * Creates a new listener instance with the given route registry.
-     *
-     * @param deploymentConfiguration
-     *            the runtime deployment configuration
-     * @param exposeServerRoutesToClient
-     *            whether to expose server routes to the client
-     */
+    private ServerAndClientViewsProvider serverAndClientViewsProvider;
+
     public RouteUnifyingIndexHtmlRequestListener(
-            DeploymentConfiguration deploymentConfiguration,
-            @Nullable NavigationAccessControl accessControl,
-            @Nullable ViewAccessChecker viewAccessChecker,
-            boolean exposeServerRoutesToClient) {
-        this.deploymentConfiguration = deploymentConfiguration;
-        this.accessControl = accessControl;
-        this.viewAccessChecker = viewAccessChecker;
-        this.exposeServerRoutesToClient = exposeServerRoutesToClient;
-
-        mapper.addMixIn(AvailableViewInfo.class, IgnoreMixin.class);
+            ServerAndClientViewsProvider serverAndClientViewsProvider) {
+        this.serverAndClientViewsProvider = serverAndClientViewsProvider;
     }
 
     @Override
     public void modifyIndexHtmlResponse(IndexHtmlResponse response) {
-        final Map<String, AvailableViewInfo> availableViews = new HashMap<>(
-                collectClientViews(response.getVaadinRequest()));
-        final boolean hasMainMenuRoute = hasMainMenu(
-                MenuRegistry.collectClientMenuItems(true,
-                        deploymentConfiguration, response.getVaadinRequest()));
-        if (exposeServerRoutesToClient) {
-            LOGGER.debug(
-                    "Exposing server-side views to the client based on user configuration");
-            availableViews.putAll(collectServerViews(hasMainMenuRoute));
-        }
-
-        if (availableViews.isEmpty()) {
-            LOGGER.debug(
-                    "No server-side nor client-side views found, skipping response modification.");
-            return;
-        }
         try {
-            final String fileRoutesJson = mapper
-                    .writeValueAsString(availableViews);
-            final String script = SCRIPT_STRING.formatted(fileRoutesJson);
+            final String script = SCRIPT_STRING
+                    .formatted(serverAndClientViewsProvider
+                            .createFileRoutesJson(response.getVaadinRequest()));
             response.getDocument().head().appendElement("script")
                     .appendChild(new DataNode(script));
         } catch (IOException e) {
@@ -121,123 +88,4 @@ public class RouteUnifyingIndexHtmlRequestListener
         }
     }
 
-    protected Map<String, AvailableViewInfo> collectClientViews(
-            VaadinRequest request) {
-
-        final Map<String, AvailableViewInfo> viewInfoMap = MenuRegistry
-                .collectClientMenuItems(true, deploymentConfiguration, request);
-
-        final Set<String> clientViewEntries = new HashSet<>(
-                viewInfoMap.keySet());
-        for (var path : clientViewEntries) {
-            if (!viewInfoMap.containsKey(path)) {
-                continue;
-            }
-
-            var viewInfo = viewInfoMap.get(path);
-            // Remove routes with required parameters, including nested ones
-            if (hasRequiredParameter(viewInfo)) {
-                viewInfoMap.remove(path);
-                if (viewInfo.children() != null) {
-                    RouteUtil.removeChildren(viewInfoMap, viewInfo, path);
-                }
-                continue;
-            }
-
-            // Remove layouts
-            if (viewInfo.children() != null) {
-                viewInfoMap.remove(path);
-            }
-        }
-
-        return viewInfoMap;
-    }
-
-    private boolean hasRequiredParameter(AvailableViewInfo viewInfo) {
-        final Map<String, RouteParamType> routeParameters = viewInfo
-                .routeParameters();
-        if (routeParameters != null && !routeParameters.isEmpty()
-                && routeParameters.values().stream().anyMatch(
-                        paramType -> paramType == RouteParamType.REQUIRED)) {
-            return true;
-        }
-
-        // Nested routes could have parameters on the parent, check them also
-        final AvailableViewInfo parentViewInfo = null;
-        if (parentViewInfo != null) {
-            return hasRequiredParameter(parentViewInfo);
-        }
-
-        return false;
-    }
-
-    protected Map<String, AvailableViewInfo> collectServerViews(
-            boolean hasMainMenu) {
-        final var vaadinService = VaadinService.getCurrent();
-        if (vaadinService == null) {
-            LOGGER.debug(
-                    "No VaadinService found, skipping server view collection");
-            return Collections.emptyMap();
-        }
-        final var serverRouteRegistry = vaadinService.getRouter().getRegistry();
-
-        var accessControls = Stream.of(accessControl, viewAccessChecker)
-                .filter(Objects::nonNull).toList();
-
-        var serverRoutes = new HashMap<String, AvailableViewInfo>();
-
-        if (vaadinService.getInstantiator().getMenuAccessControl()
-                .getPopulateClientSideMenu() == MenuAccessControl.PopulateClientMenu.ALWAYS
-                || hasMainMenu) {
-            MenuRegistry.collectAndAddServerMenuItems(
-                    RouteConfiguration.forRegistry(serverRouteRegistry),
-                    accessControls, serverRoutes);
-        }
-
-        return serverRoutes.values().stream()
-                .filter(view -> view.routeParameters().values().stream()
-                        .noneMatch(param -> param == RouteParamType.REQUIRED))
-                .collect(Collectors.toMap(this::getMenuLink,
-                        Function.identity()));
-    }
-
-    private boolean hasMainMenu(Map<String, AvailableViewInfo> availableViews) {
-        Map<String, AvailableViewInfo> clientItems = new HashMap<>(
-                availableViews);
-
-        Set<String> clientEntries = new HashSet<>(clientItems.keySet());
-        for (String key : clientEntries) {
-            if (!clientItems.containsKey(key)) {
-                continue;
-            }
-            AvailableViewInfo viewInfo = clientItems.get(key);
-            if (viewInfo.children() != null) {
-                RouteUtil.removeChildren(clientItems, viewInfo, key);
-            }
-        }
-        return !clientItems.isEmpty() && clientItems.size() == 1
-                && clientItems.values().iterator().next().route().equals("");
-    }
-
-    /**
-     * Gets menu link with omitted route parameters.
-     *
-     * @param info
-     *            the menu item's target view
-     * @return target path for menu link
-     */
-    private String getMenuLink(AvailableViewInfo info) {
-        final var parameterNames = info.routeParameters().keySet();
-        return Stream.of(info.route().split("/"))
-                .filter(Predicate.not(parameterNames::contains))
-                .collect(Collectors.joining("/"));
-    }
-
-    /**
-     * Mixin to ignore unwanted fields in the json results.
-     */
-    abstract static class IgnoreMixin {
-        @JsonIgnore
-        abstract List<AvailableViewInfo> children(); // we don't need it!
-    }
 }
